raspberry_pi/requirements.txt
# Mirqab Raspberry Pi Requirements
# Install with: pip install -r requirements.txt

# Pi Camera
picamera2>=0.3.12

# Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# HTTP Requests
requests>=2.31.0

# Optional: For advanced detection (if running model on Pi)
# torch>=2.0.0  # CPU version only on Pi
# torchvision>=0.15.0
